requirements.txt
requests~=2.32.5
pydantic~=2.11.7
google-cloud-bigquery~=3.36.0
pandas~=2.3.2
db-dtypes~=1.4.3
python-dotenv~=1.1.1
snowflake-sqlalchemy~=1.7.6
# Allow above to handle the two below
#sqlalchemy~=2.0.43
#snowflake-connector-python~=3.17.2